+++ Frontend/app/forms/models/input.model.ts
@@ -1,5 +1,0 @@
-export interface InputState {
-	past: string[];
-	present: string;
-	future: string[];
-}