+++ Frontend/app/main/components/home.component.ts
@@ -5,6 +5,7 @@ import {Component} from '@angular/core';
 	template:`
 	<h4>a CRM <small>module</small></h4>
 	<p>Welcome to Angular Bros. CRM. A place to create, send and store price quotes to companies and their related contacts.</p>
+	<quotes-component></quotes-component>
 	`
 })