+++ Frontend/app/shared/pipes/toTitleCase.pipe.ts
@@ -5,6 +5,7 @@ import {Pipe, PipeTransform} from '@angular/core';
 })
 export class Capitalize implements PipeTransform {
 	public transform(value: string, args: string[]): any {
+		console.log('hit capitalized');
 		return value.charAt(0).toUpperCase() + value.slice(1)
 	}
 }