+++ Frontend/webpack.config.js
@@ -37,7 +37,10 @@ var webpackConfig = {
 				'angular2-router-loader'
 			]
 		},
-			{ test: /\.scss$/, exclude: /node_modules/, loaders: ["style-loader", "css-loader", "sass-loader"]},
+			{
+				test: /\.scss$/,
+				exclude: /node_modules/,
+				loaders: ["style-loader", "css-loader", "sass-loader"]},
 			{ test: /\.css$/, loaders: ["to-string-loader", "css-loader"]},
 			{ test: /\.html$/, loader: "raw-loader" },